Move loading message into its own component #48
